永雏塔菲小助手，初版后端代码

+++ src/main/java/com/kanodays88/skytakeoutai/entity/querys/DishQuery.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import org.springframework.ai.tool.annotation.Tool;
 import org.springframework.ai.tool.annotation.ToolParam;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -13,9 +14,9 @@ public class DishQuery {
     @ToolParam(required = false,description = "菜品的分类类别,这个不是菜品口味，是菜品的所属分类，绝对不要往这里放入口味信息")
     private List<String> category;
     @ToolParam(required = false,description = "用户可接受的最便宜的菜品价格")
-    private Double minPrice;
+    private BigDecimal minPrice;
     @ToolParam(required = false,description = "用户可接受的最贵的菜品价格")
-    private Double maxPrice;
+    private BigDecimal maxPrice;
     @ToolParam(required = false,description = "多个菜品的名称")
     private List<String> dishNames;
 }
