+++ src/main/java/com/kanodays88/skytakeoutai/entity/querys/SetmealQuery.java
@@ -3,6 +3,7 @@ package com.kanodays88.skytakeoutai.entity.querys;
 import lombok.Data;
 import org.springframework.ai.tool.annotation.ToolParam;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -12,9 +13,9 @@ public class SetmealQuery {
     @ToolParam(required = false,description = "套餐的分类类别")
     private List<String> category;
     @ToolParam(required = false,description = "用户可接受的最便宜的菜品价格")
-    private Double minPrice;
+    private BigDecimal minPrice;
     @ToolParam(required = false,description = "用户可接受的最贵的菜品价格")
-    private Double maxPrice;
+    private BigDecimal maxPrice;
     @ToolParam(required = false,description = "多个菜品的名称")
     private List<String> dishNames;
 }
